configure Cloudinary service

+++ src/utils/cloudinary.js
@@ -2,25 +2,26 @@ import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
 cloudinary.config({ 
-    cloud_name: 'process.env.CLOUDINARY_CLOUD_NAME', 
-    api_key: 'process.env.CLOUDINARY_API_KEY', 
-    api_secret: 'process.env.CLOUDINARY_API_SECRET' // Click 'View API Keys' above to copy your API secret
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+    api_key: process.env.CLOUDINARY_API_KEY, 
+    api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
-const uploadOnCloudinary= async(localFilePath) => {
+const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(localFilePath) return null
-        //uplaod the file on cloudinary
-        const response= await cloudinary.uploader.upload(localFilePath, {
+        if (!localFilePath) return null;
+        const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
-        })
-        //file has been uploaded
-        console.log("file ahs been uplaoded on cloudinary", response.url);
-        return response
+        });
+        // console.log("file has been uploaded on cloudinary", response.url);
+        fs.unlinkSync(localFilePath)
+        return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the uplaod operation got failed 
+        console.error("Cloudinary upload error:", error);
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
         return null;
     }
 }
+
 
 export {uploadOnCloudinary}
